add functional
added the ability to change the value via input
