+++ temperature_control/src/components/input/input.css
@@ -1,0 +1,8 @@
+.input {
+  width: 100px;
+  height: 50px;
+  border: 1px solid;
+  border-radius: 10px;
+  text-align: center;
+  font-size: 35px;
+}
