+++ temperature_control/src/components/input/input.jsx
@@ -1,0 +1,7 @@
+import './input.css';
+
+function Input({ handleValue }) {
+  return <input className="input" onChange={handleValue} />;
+}
+
+export default Input;
